chore: push Oto Servis project

+++ apps/web/__tests__/finance.test.ts
@@ -21,9 +21,9 @@ describe("Property 1: Ödeme ekleme round-trip", () => {
   it("ödeme sonrası paidAmount doğru artmalı", () => {
     fc.assert(
       fc.property(
-        fc.float({ min: 0, max: 10000, noNaN: true }),
-        fc.float({ min: 100, max: 50000, noNaN: true }),
-        fc.float({ min: 0.01, max: 1000, noNaN: true }),
+        fc.double({ min: 0, max: 10000, noNaN: true }),
+        fc.double({ min: 100, max: 50000, noNaN: true }),
+        fc.double({ min: 0.01, max: 1000, noNaN: true }),
         (currentPaid, totalAmount, paymentAmount) => {
           fc.pre(currentPaid < totalAmount);
           fc.pre(currentPaid + paymentAmount <= totalAmount);
@@ -38,7 +38,7 @@ describe("Property 1: Ödeme ekleme round-trip", () => {
   it("paidAmount >= totalAmount olduğunda status PAID olmalı", () => {
     fc.assert(
       fc.property(
-        fc.float({ min: 100, max: 50000, noNaN: true }),
+        fc.double({ min: 100, max: 50000, noNaN: true }),
         (totalAmount) => {
           const result = simulatePayment(0, totalAmount, totalAmount);
           return result.newStatus === "PAID";
@@ -51,8 +51,8 @@ describe("Property 1: Ödeme ekleme round-trip", () => {
   it("paidAmount < totalAmount olduğunda status SENT kalmalı", () => {
     fc.assert(
       fc.property(
-        fc.float({ min: 100, max: 50000, noNaN: true }),
-        fc.float({ min: 0.01, max: 0.99, noNaN: true }),
+        fc.double({ min: 100, max: 50000, noNaN: true }),
+        fc.double({ min: 0.01, max: 0.99, noNaN: true }),
         (totalAmount, ratio) => {
           const partialPayment = totalAmount * ratio;
           const result = simulatePayment(0, totalAmount, partialPayment);
